refactor(common): move get_anthias_release into anthias_common.version
``anthias_common.http`` previously reached up into
``anthias_server.lib.diagnostics`` for the release string, inverting
the layering ``anthias_common`` is supposed to have (it's already
imported by both ``anthias_server`` and ``anthias_viewer``; the
viewer would otherwise drag in Django via diagnostics if it ever
adopted ``AnthiasSession``).

Move the pure version-reader (importlib.metadata + pyproject.toml
fallback, both stdlib) to ``anthias_common.version`` and re-export
from the old diagnostics path so existing call sites keep working.

Per #2897 Copilot review.
