+++ src/anthias_server/lib/diagnostics.py
@@ -7,6 +7,12 @@ from datetime import datetime
 
 from anthias_common import device_helper, utils
 
+# Re-exported from ``anthias_common.version`` (the ``as`` form marks
+# this as an explicit re-export so it stays importable from the old
+# diagnostics path without a lint suppression). Layer-agnostic code
+# imports it from ``anthias_common.version`` directly.
+from anthias_common.version import get_anthias_release as get_anthias_release
+
 
 _CEC_QUERY_SCRIPT = """
 import sys
@@ -189,60 +195,6 @@ def get_git_hash() -> str | None:
 _RELEASE_BRANCHES = frozenset({'master', 'main'})
 
 
-def get_anthias_release() -> str:
-    """Read the project version, sourced from pyproject.toml's
-    [project].version (currently CalVer ``YYYY.M.MICRO``).
-
-    Resolution order:
-      1. ``importlib.metadata.version('anthias')`` — works for
-         editable installs (``pip install -e .``) and any path where
-         the project ships as a wheel.
-      2. Fallback: parse ``pyproject.toml`` directly with ``tomllib``.
-         Required because every production / test / host environment
-         runs ``uv sync --no-install-project`` (see
-         docker/uv-builder.j2, docker/Dockerfile.{server,test,viewer},
-         bin/install.sh) — that flag installs the project's deps but
-         NOT the project itself, so importlib.metadata has no record
-         of an ``anthias`` distribution to read.
-
-    Cached after first successful read so the System Info HTML render
-    (called per request) and the v2 info API don't re-open the file.
-    Returns the empty string only when both sources fail.
-    """
-    cached: str | None = getattr(get_anthias_release, '_cached', None)
-    if cached is not None:
-        return cached
-    value: str
-    try:
-        from importlib.metadata import PackageNotFoundError, version
-
-        try:
-            value = version('anthias')
-        except PackageNotFoundError:
-            value = _read_version_from_pyproject()
-    except Exception:
-        value = ''
-    setattr(get_anthias_release, '_cached', value)
-    return value
-
-
-def _read_version_from_pyproject() -> str:
-    """Last-ditch source for the project version when the package
-    isn't installed in the active venv. Walks up from this file to
-    find the repo-root pyproject.toml — `__file__` lives at
-    ``src/anthias_server/lib/diagnostics.py``, so parents[3] is the
-    checkout root in both editable installs and the
-    ``uv sync --no-install-project`` Docker layout."""
-    try:
-        import tomllib
-        from pathlib import Path
-
-        pyproject = Path(__file__).resolve().parents[3] / 'pyproject.toml'
-        with pyproject.open('rb') as f:
-            data = tomllib.load(f)
-        return str(data.get('project', {}).get('version', ''))
-    except Exception:
-        return ''
 
 
 def get_anthias_version_head() -> str:
